# Add caching in iterateStackTrace