Add a global (VM thread shared) cache mapping JIT method PCs to data needed for the callback in `iterateStackTrace`. Since this cache is global and multiple VM threads can be reading/writing we also add locking around the cache.

Issue: #7776

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>