@fjeremic 2019-12-10
In terms of performance I've ran this on a large webserver workload which exercises this path in the profile and it showed a roughly 2% improvement. In terms of smaller workloads which are more isolated I have a unit benchmark [1] which hammers away at backtrace generation. This change gives roughly 3x improvement in throughput performance:

Baseline:
```
# Run progress: 0.00% complete, ETA 00:00:40
# Fork: 1 of 1
# Warmup Iteration   1: 14448.300 ops/s
Iteration   1: 15558.908 ops/s
Iteration   2: 16550.354 ops/s
Iteration   3: 15340.605 ops/s


Result "org.openj9.GetStackTraceBenchmark.testgetStackTrace":
  15816.622 ±(99.9%) 11762.400 ops/s [Average]
  (min, avg, max) = (15340.605, 15816.622, 16550.354), stdev = 644.737
  CI (99.9%): [4054.223, 27579.022] (assumes normal distribution)
```

Proposed change:
```
# Run progress: 0.00% complete, ETA 00:00:40
# Fork: 1 of 1
# Warmup Iteration   1: 37428.256 ops/s
Iteration   1: 48420.139 ops/s
Iteration   2: 45614.409 ops/s
Iteration   3: 52355.780 ops/s


Result "org.openj9.GetStackTraceBenchmark.testgetStackTrace":
  48796.776 ±(99.9%) 61781.108 ops/s [Average]
  (min, avg, max) = (45614.409, 48796.776, 52355.780), stdev = 3386.431
  CI (99.9%): [≈ 0, 110577.885] (assumes normal distribution)
```

This was measured on a 4-core x86 machine using the following JVM:

```
openjdk version "11.0.5" 2019-10-15
OpenJDK Runtime Environment AdoptOpenJDK (build 11.0.5+10-201910301851)
Eclipse OpenJ9 VM AdoptOpenJDK (build master-99e396a57, JRE 11 Linux amd64-64-Bit Compressed References 20191030_359 (JIT enabled, AOT enabled)
OpenJ9   - 99e396a57
OMR      - cf9d75a4
JCL      - cc4272e86c based on jdk-11.0.5+10)
```

[1] https://github.com/fjeremic/jmh-benchmarks