@DanHeidinga 2020-01-13
@fjeremic Do you have any data on which values are the most expensive to lookup?  Or put another way, which are the most valuable to cache?

There's very similar logic in jswalk.c which has a per-thread cache stored in the `J9VMThread->jitExceptionHandlerCache` field to optimize mapping `jitPC` to `J9JITExceptionTable`.  Was this code based on that code?
https://github.com/eclipse/openj9/blob/6b1f6d9b5e49a7797c2beb8cb51df3c4a0b602b1/runtime/codert_vm/jswalk.c#L1483

This would be the second `jitPC` -> `J9JITExceptionTable` cache (though this one caches additional fields as well) that makes me wonder if there's a bigger problem lurking in how we do the mapping from pc to metadata.  Would we get more benefit by looking at that code rather than special casing here?

The cache will serialize multiple threads that are generating stacktraces as there is a single global cache.  I don't see a case showing the potential cache contention in the benchmark results.

Once we've worked though the above points, we'll also need to add cache-cleanup code to handle removing stale entries when class loaders get unloaded.  It's possible that a new PC could be found in the cache pointing to metadata for an unloaded class.