@fjeremic 2020-01-13
> Once we've worked though the above points, we'll also need to add cache-cleanup code to handle removing stale entries when class loaders get unloaded. It's possible that a new PC could be found in the cache pointing to metadata for an unloaded class.

Thinking about this is it possible for us to have a method on the backtrace of a class while the respective class is unloaded? I would have imagined if we are executing a method, i.e. it is somewhere on the backtrace then the cache entries for all the methods on the backtrace should be valid.