@DanHeidinga 2020-01-14
> If getting a stale entry is possible would you be able to explain the scenario in which it could occur? 

When an exception is created, it walks the stack and collects an array of PCs for the methods currently executing.  This array of PCs is not a gc root and doesn't keep the classes alive.

If the application holds onto the exception and only prints it later after a GC has unloaded classes, some of the PCs may no longer be valid as the classes they referred to have been unloaded.

This is why the cache in `jswalk` is cleaned up in on of the gc-related hooks.