@DanHeidinga 2020-01-14
Thanks for sharing the data @fjeremic. 

> `   + 9.08% findROMMethodInROMClass`

I think this can be sped in another way.  ROMMethods are variable sized making walking them more expensive.  For JITTed methods, we have the RAMClass in hand already which would allow walking the `J9Methods` which are fixed sized.  This should improve the speed of that operation.

> `+ 6.10% peekClassHashTable`

This can also be sped up or removed in a lot of cases by changing the iterator callback to optionally take the RAMClass as a parameter.  In the JITTed method case, the lookup won't be required anymore as we already have the RAMClass in the iterator and as per above, the interpreter case will have fetched it as well to speed the method walk.

I'd like to see how far those changes take us rather than adding the global lock and the cache