@fjeremic 2020-01-29
> What is the next step on this ?

Still paging on cycles to get back to this. I will give a shot to what Dan has suggest to see if we can get similar performance.