@fjeremic 2020-02-03
I'm prototyping the change to `getStackTraceIterator` to pass in the `J9Class` pointer optionally if available and avoid the calls to `peekClassHashTable` now. Will post results when I have them.