@fjeremic 2020-02-04
I've prototyped the two changes proposed. Here are the performance results on the JMH benchmarks linked above:

| Benchmark         | Baseline | Cache | Peek  | Peek + J9Class | 
|-------------------|----------|-------|-------|----------------| 
| JMH Single-Thread | 37746    | 72290 | 15632 | 16107          | 
| JMH Multi-Thread  | 69076    | 81794 | 28978 | 30174          | 


`Cache` corresponds to the caching proposed in this PR, `Peek` corresponds to the improvements around changing the iterator callback to optionally take the RAMClass as a parameter (https://github.com/fjeremic/openj9/commit/edfe8b0b5a4d4ee08310d59cecf62e50e9ff142c), and `Peek + J9Class` is the previous change in addition to the change to walk `J9Class` instead of `J9ROMClass` as outlined in @DanHeidinga's review (https://github.com/fjeremic/openj9/commit/8f7a6a1f028cdfc63cde44e74b69d57605a7f9d1).

The results indicate that with the current prototypes we cannot get close to the performance which caching gives to us. It seems to actually slow us down, at least on the JMH benchmark. I tried running both with `-Xint` which showed roughly on-par behaviour and `-Xjit:count=1` which showed a regression on the two prototypes (Peek and Peek + J9Class).

Please advise how we should proceed.